openshift compatibility fixes

+++ testsuite/utils/jpa/backuprestore.go
@@ -3,7 +3,6 @@ package jpa
 import (
 	"context"
 	"net/http"
-	"os"
 	"strconv"
 	"time"
 
@@ -52,40 +51,35 @@ func ExecuteBackupAndRestoreTestCase(suiteCtx *suite.SuiteContext, ctx *types.Te
 			},
 		},
 	}
-	apicurioutils.CreateRegistryAndWait(suiteCtx, ctx, &backupregistry)
-	functional.BasicRegistryAPITest(ctx)
 	ctx.RegisterCleanup(func() {
 		if apicurioutils.ExistsRegistry(suiteCtx, backupregistry.Name) {
 			apicurioutils.DeleteRegistryAndWait(suiteCtx, backupregistry.Name)
 		}
 	})
+	apicurioutils.CreateRegistryAndWait(suiteCtx, ctx, &backupregistry)
+	functional.BasicRegistryAPITest(ctx)
 
 	//create artifacts on the registry
-	client := apicurioclient.NewApicurioRegistryApiClient(ctx.RegistryHost, ctx.RegistryPort, http.DefaultClient)
+	backupclient := apicurioclient.NewApicurioRegistryApiClient(ctx.RegistryHost, ctx.RegistryPort, http.DefaultClient)
 
 	for i := 1; i <= 50; i++ {
-		err := client.CreateArtifact("bandr-"+strconv.Itoa(i), apicurioclient.Avro, artifactData)
+		err := backupclient.CreateArtifact("bandr-"+strconv.Itoa(i), apicurioclient.Avro, artifactData)
 		Expect(err).ToNot(HaveOccurred())
 		time.Sleep(1 * time.Second)
 	}
 
-	artifacts, err := client.ListArtifacts()
+	artifacts, err := backupclient.ListArtifacts()
 	Expect(err).ToNot(HaveOccurred())
 	Expect(len(artifacts)).To(BeIdenticalTo(50))
 
 	// deploy a dummypod to create the backup and store it, and then restore the backup from that pod
 	log.Info("Deploying dbplayground")
 
-	oldDir, err := os.Getwd()
-	Expect(err).NotTo(HaveOccurred())
-	dbplaygroundBuildDir := utils.SuiteProjectDirValue + "/scripts/dbplayground"
-	os.Chdir(dbplaygroundBuildDir)
-	err = utils.ExecuteCmd(true, &utils.Command{Cmd: []string{"make", "build", "push"}})
-	os.Chdir(oldDir)
-
-	//TODO make this work on openshift
-	dbplaygroundImage := "localhost:5000/dbplayground:latest"
-
+	dbplaygroundImage := "quay.io/rh_integration/service-registry-dbplayground:pg10"
+	kubernetescli.Execute("create", "serviceaccount", "dbplayground", "-n", utils.OperatorNamespace)
+	if suiteCtx.IsOpenshift {
+		kubernetescli.Execute("adm", "policy", "add-scc-to-user", "privileged", "system:serviceaccount:"+utils.OperatorNamespace+":dbplayground", "-n", utils.OperatorNamespace)
+	}
 	err = suiteCtx.K8sClient.Create(context.TODO(), dbplaygroundDeployment(dbplaygroundImage))
 	Expect(err).ToNot(HaveOccurred())
 	kubernetesutils.WaitForDeploymentReady(suiteCtx.Clientset, 120*time.Second, "dbplayground", 1)
@@ -94,6 +88,9 @@ func ExecuteBackupAndRestoreTestCase(suiteCtx *suite.SuiteContext, ctx *types.Te
 	podList, err := suiteCtx.Clientset.CoreV1().Pods(utils.OperatorNamespace).List(metav1.ListOptions{LabelSelector: labelsSet.AsSelector().String()})
 	Expect(err).ToNot(HaveOccurred())
 	dbplaygroundPodName := podList.Items[0].Name
+	ctx.RegisterCleanup(func() {
+		suiteCtx.Clientset.AppsV1().Deployments(utils.OperatorNamespace).Delete("dbplayground", &metav1.DeleteOptions{})
+	})
 
 	// create the backup
 	kubernetescli.Execute("-n", utils.OperatorNamespace, "exec", dbplaygroundPodName, "--", "./create_backup.sh", backupDBData.Host, backupDBData.Port, backupDBData.Database, backupDBData.User, backupDBData.Password)
@@ -131,14 +128,17 @@ func ExecuteBackupAndRestoreTestCase(suiteCtx *suite.SuiteContext, ctx *types.Te
 			},
 		},
 	}
+	ctx.RegisterCleanup(func() {
+		if apicurioutils.ExistsRegistry(suiteCtx, restoreregistry.Name) {
+			apicurioutils.DeleteRegistryAndWait(suiteCtx, restoreregistry.Name)
+		}
+	})
 	apicurioutils.CreateRegistryAndWait(suiteCtx, ctx, &restoreregistry)
 	functional.BasicRegistryAPITest(ctx)
-	ctx.RegisterCleanup(func() {
-		apicurioutils.DeleteRegistryAndWait(suiteCtx, restoreregistry.Name)
-	})
 
 	// verify new registry have old data
-	artifacts, err = client.ListArtifacts()
+	restoreclient := apicurioclient.NewApicurioRegistryApiClient(ctx.RegistryHost, ctx.RegistryPort, http.DefaultClient)
+	artifacts, err = restoreclient.ListArtifacts()
 	Expect(err).ToNot(HaveOccurred())
 	Expect(len(artifacts)).To(BeIdenticalTo(50))
 
@@ -146,6 +146,7 @@ func ExecuteBackupAndRestoreTestCase(suiteCtx *suite.SuiteContext, ctx *types.Te
 
 func dbplaygroundDeployment(image string) *v1.Deployment {
 	var replicas int32 = 1
+	var privileged bool = true
 	return &v1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "dbplayground",
@@ -161,10 +162,14 @@ func dbplaygroundDeployment(image string) *v1.Deployment {
 					Labels: dbplaygroundlabels,
 				},
 				Spec: corev1.PodSpec{
+					ServiceAccountName: "dbplayground",
 					Containers: []corev1.Container{
 						{
 							Name:  "dbplayground",
 							Image: image,
+							SecurityContext: &corev1.SecurityContext{
+								Privileged: &privileged,
+							},
 						},
 					},
 				},
